suiclient/object_owner.go
package suiclient

import (
	"bytes"
	"encoding/json"
	"errors"

	"github.com/pattonkan/sui-go/sui"
)

type ObjectOwnerInternal struct {
	AddressOwner *sui.Address `json:"AddressOwner,omitempty"`
	ObjectOwner  *sui.Address `json:"ObjectOwner,omitempty"`
	SingleOwner  *sui.Address `json:"SingleOwner,omitempty"`
	Shared       *struct {
		InitialSharedVersion *sui.SequenceNumber `json:"initial_shared_version"`
	} `json:"Shared,omitempty"`
}

type ObjectOwner struct {
	*ObjectOwnerInternal
	*string
}

func (o ObjectOwner) MarshalJSON() ([]byte, error) {
	if o.string != nil {
		data, err := json.Marshal(o.string)
		if err != nil {
			return nil, err
		}
		return data, nil
	}
	if o.ObjectOwnerInternal != nil {
		data, err := json.Marshal(o.ObjectOwnerInternal)
		if err != nil {
			return nil, err
		}
		return data, nil
	}
	return nil, errors.New("nil value")
}

func (o *ObjectOwner) UnmarshalJSON(data []byte) error {
	if bytes.HasPrefix(data, []byte("\"")) {
		stringData := string(data[1 : len(data)-1])
		o.string = &stringData
		return nil
	}
	if bytes.HasPrefix(data, []byte("{")) {
		oOI := ObjectOwnerInternal{}
		err := json.Unmarshal(data, &oOI)
		if err != nil {
			return err
		}
		o.ObjectOwnerInternal = &oOI
		return nil
	}
	return errors.New("value not json")
}
